// src/app/core/models/prestadores/prestador.model.ts
export class Prestadores {
    id: string;
    name: string;
    contact: string;
    state: string;
    city: string;
    cpf: string;
    birth: string;
    email: string;
    stateSigla: string;
    image: string;
}

export class Services {
    id: string;
    name: string;
    uc: string;
    description: string;
    email: string;
}